Fix: Service Worker for notification
